optimisation

+++ capteur_analog.cpp
@@ -1,6 +1,6 @@
 #include "capteur_analog.h"
 
-CapteurAnalog::CapteurAnalog(byte id, int pin) : Capteur::Capteur(id, 1){
+CapteurAnalog::CapteurAnalog(byte id, byte pin) : Capteur::Capteur(id, 1){
   _pin = pin;
 }
