+++ capteur_analog.h
@@ -13,11 +13,11 @@
 
 class CapteurAnalog : public Capteur{
 public:
-	CapteurAnalog(byte id, int pin); // Declaration du capteur avec sa pin
+	CapteurAnalog(byte id, byte pin); // Declaration du capteur avec sa pin
         virtual bool refresh();
         virtual bool init();
 protected:
-        int _pin; // Patte sur laquelle le capteur est connecté
+        byte _pin; // Patte sur laquelle le capteur est connecté
 };
 
 #endif
